Rename ObjectData.hpp -> ModelData.hpp
